requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
face-recognition
Pillow
numpy
pytest
httpx
